Move ConfigurationFileUtils to .test.util
