+++ plugins/com.google.cloud.tools.eclipse.appengine.facets.test/src/com/google/cloud/tools/eclipse/appengine/facets/ui/navigator/model/ModelRefreshTests.java
@@ -26,7 +26,7 @@ import static org.junit.Assert.assertTrue;
 
 import com.google.cloud.tools.appengine.api.AppEngineException;
 import com.google.cloud.tools.eclipse.appengine.facets.AppEngineStandardFacet;
-import com.google.cloud.tools.eclipse.appengine.facets.ui.navigator.ConfigurationFileUtils;
+import com.google.cloud.tools.eclipse.test.util.project.ConfigurationFileUtils;
 import com.google.cloud.tools.eclipse.test.util.project.ProjectUtils;
 import com.google.cloud.tools.eclipse.test.util.project.TestProjectCreator;
 import com.google.cloud.tools.eclipse.util.io.ResourceUtils;
@@ -183,7 +183,7 @@ public class ModelRefreshTests {
     ConfigurationFileUtils.createEmptyQueueXml(project);
     AppEngineStandardProjectElement projectElement =
         AppEngineStandardProjectElement.create(project);
-    final AppEngineResourceElement[] subElements = projectElement.getConfigurations();
+    AppEngineResourceElement[] subElements = projectElement.getConfigurations();
     assertEquals(5, subElements.length);
     assertThat(subElements, hasItemInArray(instanceOf(CronDescriptor.class)));
     assertThat(subElements, hasItemInArray(instanceOf(DatastoreIndexesDescriptor.class)));
@@ -215,7 +215,7 @@ public class ModelRefreshTests {
     ConfigurationFileUtils.createEmptyQueueXml(project);
     AppEngineStandardProjectElement projectElement =
         AppEngineStandardProjectElement.create(project);
-    final AppEngineResourceElement[] subElements = projectElement.getConfigurations();
+    AppEngineResourceElement[] subElements = projectElement.getConfigurations();
     assertEquals(5, subElements.length);
     assertThat(subElements, hasItemInArray(instanceOf(CronDescriptor.class)));
     assertThat(subElements, hasItemInArray(instanceOf(DatastoreIndexesDescriptor.class)));
@@ -248,7 +248,7 @@ public class ModelRefreshTests {
     AppEngineStandardProjectElement projectElement =
         AppEngineStandardProjectElement.create(project);
     files.add(projectElement.getDescriptorFile());
-    final AppEngineResourceElement[] subElements = projectElement.getConfigurations();
+    AppEngineResourceElement[] subElements = projectElement.getConfigurations();
 
     for (IFile file : files) {
       boolean changed = projectElement.resourcesChanged(Collections.singleton(file));
@@ -291,15 +291,15 @@ public class ModelRefreshTests {
     IFile oldCronXml = ConfigurationFileUtils.createEmptyCronXml(project);
     AppEngineStandardProjectElement projectElement =
         AppEngineStandardProjectElement.create(project);
-    final AppEngineResourceElement[] oldElements = projectElement.getConfigurations();
+    AppEngineResourceElement[] oldElements = projectElement.getConfigurations();
     assertEquals(1, oldElements.length);
     assertThat(oldElements, hasItemInArray(instanceOf(CronDescriptor.class)));
 
     // create the new WEB-INF location and populate it
-    final IFolder newWebRoot = project.getFolder("newWebRoot");
-    final IFolder newWebInf = newWebRoot.getFolder("WEB-INF");
+    IFolder newWebRoot = project.getFolder("newWebRoot");
+    IFolder newWebInf = newWebRoot.getFolder("WEB-INF");
     ResourceUtils.createFolders(newWebInf, null);
-    final IFile newDispatchXml = newWebInf.getFile("dispatch.xml");
+    IFile newDispatchXml = newWebInf.getFile("dispatch.xml");
     newDispatchXml.create(
         new ByteArrayInputStream("<dispatch-entries/>".getBytes(StandardCharsets.UTF_8)),
         true,
@@ -307,8 +307,8 @@ public class ModelRefreshTests {
     assertTrue("error creating new dispatch.xml", newDispatchXml.exists());
 
     // now link in the new WEB-INF into the overlay
-    final IWorkspace workspace = project.getWorkspace();
-    final Set<IFile> changed =
+    IWorkspace workspace = project.getWorkspace();
+    Set<IFile> changed =
         recordChangedFilesDuring(
             workspace,
             monitor -> {
@@ -325,7 +325,7 @@ public class ModelRefreshTests {
         Iterables.getOnlyElement(changed));
     assertTrue(projectElement.resourcesChanged(changed));
 
-    final AppEngineResourceElement[] newElements = projectElement.getConfigurations();
+    AppEngineResourceElement[] newElements = projectElement.getConfigurations();
     assertEquals(2, newElements.length);
     assertThat(newElements, hasItemInArray(instanceOf(CronDescriptor.class)));
     assertThat(newElements, hasItemInArray(instanceOf(DispatchRoutingDescriptor.class)));
@@ -339,8 +339,8 @@ public class ModelRefreshTests {
   /** Record and return the set of files altered when running the provided block. */
   private static Set<IFile> recordChangedFilesDuring(IWorkspace workspace, ICoreRunnable block)
       throws CoreException {
-    final Set<IFile> changed = new LinkedHashSet<>();
-    final IResourceChangeListener listener =
+    Set<IFile> changed = new LinkedHashSet<>();
+    IResourceChangeListener listener =
         event -> {
           try {
             changed.addAll(ResourceUtils.getAffectedFiles(event.getDelta()).values());
